Created separate components for navigation and tasks
